refactor: alterar tipo do ID em Employee

+++ src/main/java/com/business_management/domain/Employee.java
@@ -17,7 +17,7 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int employee_id;
+    private Long employee_id;
 
     @Column(name = "name")
     private String name;
@@ -35,11 +35,11 @@ public class Employee {
     public Employee() {
     }
 
-    public int getEmployee_id() {
+    public Long getEmployee_id() {
         return employee_id;
     }
 
-    public void setEmployee_id(int employee_id) {
+    public void setEmployee_id(Long employee_id) {
         this.employee_id = employee_id;
     }
